+++ src/JointPaths/Finder/Finder.cpp
@@ -1,8 +1,9 @@
-#include "EntanglementFinder.hh"
+#include "Finder.hh"
 
 using namespace std;
+using namespace JointPaths;
 
-int EntanglementFinder::GetInitialLevelIfValid(vector<Path>* paths) {
+int Finder::GetInitialLevelIfValid(vector<Path>* paths) {
 	int level = 2;
 	if (Data.SearchCeiling != -1)
 		level = Data.SearchCeiling;
@@ -28,7 +29,7 @@ int EntanglementFinder::GetInitialLevelIfValid(vector<Path>* paths) {
 	return level;
 }
 
-unordered_map<size_t, EntanglementOccurance> EntanglementFinder::FindEntangledCandidates(vector<Path>* paths) {
+unordered_map<size_t, EntanglementOccurance> Finder::FindEntangledCandidates(vector<Path>* paths) {
 	unordered_map<size_t, EntanglementOccurance> candidates;
 
 	if (paths->size() == 0)
@@ -65,7 +66,7 @@ unordered_map<size_t, EntanglementOccurance> EntanglementFinder::FindEntangledCa
 	//return unordered_map<size_t, EntanglementOccurance>(candidates);
 }
 
-int EntanglementFinder::ReduceLevel(int level) {
+int Finder::ReduceLevel(int level) {
 	int newLevel = level;
 	if (Data.LevelReductionType == RunData::LevelReductionTypes::Division)
 		newLevel = ceil((double)level / Data.LevelReductionFactor);
@@ -76,7 +77,7 @@ int EntanglementFinder::ReduceLevel(int level) {
 	return newLevel;
 }
 
-void EntanglementFinder::GenerateActionSet(vector<pair<size_t, vector<PDDLActionInstance*>>>* currentValues, vector<Path>* paths, const int level) {
+void Finder::GenerateActionSet(vector<pair<size_t, vector<PDDLActionInstance*>>>* currentValues, vector<Path>* paths, const int level) {
 	const int pathsSize = paths->size();
 	for (int i = 0; i < pathsSize; i++) {
 		Path* path = &paths->at(i);
@@ -101,7 +102,7 @@ void EntanglementFinder::GenerateActionSet(vector<pair<size_t, vector<PDDLAction
 	}
 }
 
-void EntanglementFinder::AddCandidatesIfThere(unordered_map<size_t, EntanglementOccurance>* candidates, const vector<pair<size_t, vector<PDDLActionInstance*>>>* currentValues) {
+void Finder::AddCandidatesIfThere(unordered_map<size_t, EntanglementOccurance>* candidates, const vector<pair<size_t, vector<PDDLActionInstance*>>>* currentValues) {
 	const int currentValueSize = currentValues->size();
 	if (OnNewLevel != nullptr)
 		OnNewLevel(_CurrentLevel, currentValueSize);
@@ -138,7 +139,7 @@ void EntanglementFinder::AddCandidatesIfThere(unordered_map<size_t, Entanglement
 		OnLevelEnd();
 }
 
-bool EntanglementFinder::IsOverTimeLimit() {
+bool Finder::IsOverTimeLimit() {
 	if (_HasTimeLimit) {
 		auto currentTime = chrono::steady_clock::now();
 		auto ellapsed = chrono::duration_cast<chrono::milliseconds>(currentTime - _StartTime).count();
